+++ course-registration-service/src/main/java/org/example/courseregistration/repository/KeeperRejectionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface KeeperRejectionRepository extends JpaRepository<KeeperRejection, Integer> {
-    Optional<KeeperRejection> findKeeperRejectionByResponseId(Integer responseId);
+public interface KeeperRejectionRepository extends JpaRepository<KeeperRejection, Long> {
+    Optional<KeeperRejection> findKeeperRejectionByResponseId(Long responseId);
 }
